requirements.txt
gradio
openai
openai-agents
pydantic
python-dotenv